apps/api/requirements.txt
# Django core
django>=5.1,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.0
dj-database-url>=2.1
python-dotenv>=1.0

# Database
psycopg2-binary>=2.9

# Supabase
supabase>=2.0
gotrue>=2.0

# File processing (used by extraction package)
pdfplumber>=0.11
pytesseract>=0.3
Pillow>=10.0
pandas>=2.2
openpyxl>=3.1

# AI / NLP
openai>=1.30

# PDF generation
reportlab>=4.1
weasyprint>=62.0

# Task queue (optional, for async extraction)
celery>=5.3
redis>=5.0

# Testing
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3

# Utils
python-dateutil>=2.9
